start adapting

+++ gnn_agglomeration/nn/models/model_type/regression_problem.py
@@ -39,7 +39,7 @@ class RegressionProblem(ModelType):
 
     def metric(self, predictions, targets, mask):
         # TODO test this
-        weighted_equal = torch.squeeze(predictions).eq(
+        weighted_equal = torch.squeeze(predictions.float()).eq(
             targets.float()).float() * mask.float()
         correct = weighted_equal.sum().item()
         acc = correct / (mask.sum().item() + torch.finfo(torch.float).tiny)
