+++ gnn_agglomeration/pyg_datasets/toy_datasets/count_neighbors_graph.py
@@ -20,7 +20,7 @@ class CountNeighborsGraph(MyGraph):
 
         # connect all edges within distance theta_max O(n^2)
         edges = []
-        y = torch.zeros(config.nodes, dtype=torch.long)
+        y = torch.zeros(config.nodes, dtype=torch.float)
         x = torch.arange(
             config.nodes) % config.feature_dimensionality
